feat(consensus): dynamic validator sets (#1039)
This patch allows dynamically adding and removing validators from Tempo
by using the `ValidatorConfig` smart contract. The consensus layer is
hardfork enabled, meaning that it only transitions to the new behavior
after `allegrettoTime`.

There is a high level overview of the protocol provided in
`docs/network-reconfiguration.md` provided with this patch.

## Notable changes

### Configuration

1. consensus config was removed in favor of command line arguments, see
`tempo node --help` and `crates/commonware-node/src/args.rs`
2 the public polynomial was moved to chainspec `"publicPolynomial"`
entry
3. the initial set of peers was moved to chainspec `"validators"` entry
4. epoch length was moved to chainspec `"epochLength"` entry

### Adding and removing validators

1. where before `dealers == players` in each ceremony (and where both
were defined through chainspec), the players in a new ceremony are now
established by doing a `getValidators` call of the `ValidatorConfig`
smart contract (technically, the validators returned by the contract
call will be the players in the ceremony of the epoch after, giving them
time to sync for one entire epoch; see the spec provided with this
patch).
2. the p2p peers are similarly established through `getValidators`,
where we track the peers of the last 3 epochs (current, parent,
grandparent). The peers in epoch N is the union of the peers of `N`,
`N+1` and `N+2`.
3. validators are removed by using the `changeValidatorStatus` smart
contract call setting `active = false`. Similar to how adding validators
work, this will remove the validator from first the `syncing players`
list, and then from the `players` of the ceremony in the epoch after the
next.

### DKG ceremonies and hardfork behavior

1. at startup, DKG ceremonies are run with the validator set specified
in the chainspec. This matches the behavior before this patch (except
that peers were provided through a config file)
2. when a new DKG ceremony is started (on the last height of the epoch)
and if `boundary_block.timestamp >= allegretto_time`, the DKG manager
reads the entries of the `ValidatorConfig` smart contract
4. if the entries match exactly the validators in the chainspec (i.e.
`chainspec.validators == contract.validators` with both being a list
`[(pubkey:socketaddr)]`), then a DKG ceremony according to post-hardfork
logic is started.
5. from then on, validators can be added and removed by using the
`ValidatorConfig` smart contract.
6. in all other cases (timestamp is before allegretto time, the contract
validators don't match), the pre-hardfork logic stays in place until the
next epoch. This gives the operator time to fix the onchain contract
state and provide the correct validator data.

Note 1: this requires an operator (admin account) to issue explicit
`addValidator` calls to add all validators to the contract state.
Note 2: at the moment, a node always starts out in the pre-hardfork
logic even if allegretto time is 0. This will be fixed in a followup.

### Xtasks

1. update xtask genesis generator to generate and inject validators,
public polynomial into genesis/chainspec
2. rework the `generate-genesis` task by breaking out genesis-generation
into its own set of arguments
3. rework the `generate-devnet` task in terms of this this new
genesis-generation
4. remove `generate-config`

### e2e testing

1. rework e2e tests to get access to configs, execute smart contract
calls from within the execution/tokio runtime
2. provide an e2e test for adding and removing a validator, a validator
having lost its share and regaining it in a subsequent round, transition
from pre- to post-allegretto logic

## Followup issues

https://github.com/tempoxyz/tempo/issues/562
https://github.com/tempoxyz/tempo/issues/565
https://github.com/tempoxyz/tempo/issues/792
https://github.com/tempoxyz/tempo/issues/884
https://github.com/tempoxyz/tempo/issues/924
https://github.com/tempoxyz/tempo/issues/928
https://github.com/tempoxyz/tempo/issues/930
https://github.com/tempoxyz/tempo/issues/933
https://github.com/tempoxyz/tempo/issues/1087

Closes #310 
Closes #563 
Closes #883

---------

Co-authored-by: Arsenii Kulikov <klkvrr@gmail.com>
Co-authored-by: zhygis <5236121+Zygimantass@users.noreply.github.com>